Changed WvTest's equality check to use off_t instead of int.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/idb2@10838 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ include/wvtest.h
@@ -13,6 +13,7 @@
 #define __WVTEST_H
 
 #include <time.h>
+#include <sys/types.h>
 
 class WvTest
 {
@@ -37,7 +38,7 @@ public:
         { start(file, line, condstr); check(cond); return cond; }
     static bool start_check_eq(const char *file, int line,
 			       const char *a, const char *b, bool expect_pass);
-    static bool start_check_eq(const char *file, int line, int a, int b,
+    static bool start_check_eq(const char *file, int line, off_t a, off_t b,
                                bool expect_pass);
     static bool start_check_lt(const char *file, int line,
                                const char *a, const char *b);
